Initial commit for showing round values.
As part of this, I added in code that better handles setting the min and
max values: they are automatically switched in the tip manager (but not
in preferences) if the min is greater than the max. Also, if they are
equal, only show one of them.

Specifying showing round tip amounts will show each dollar increment
between the min an max. However, this has horrible performance on large
values (so may need to rethink this a bit.

I added a warning note to the option summary (although a dialog would
probably be better).